+++ backend/src/main/java/com/example/fintrack/security/TokenType.java
@@ -1,4 +1,4 @@
-package com.example.fintrack.security.enums;
+package com.example.fintrack.security;
 
 public enum TokenType {
     ACCESS,
