+++ backend/src/main/java/com/example/fintrack/security/service/RedisService.java
@@ -1,6 +1,6 @@
 package com.example.fintrack.security.service;
 
-import com.example.fintrack.security.enums.TokenType;
+import com.example.fintrack.security.TokenType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
@@ -11,9 +11,9 @@ import java.util.concurrent.TimeUnit;
 @RequiredArgsConstructor
 public class RedisService {
 
-    private final RedisTemplate<String, Object> redisTemplate;
-
     private static final long REFRESH_EXPIRATION = 30L;
+
+    private final RedisTemplate<String, Object> redisTemplate;
 
     public void saveToken(String email, String token, TokenType tokenType) {
         if (tokenType == TokenType.REFRESH) {
@@ -24,6 +24,5 @@ public class RedisService {
 
     public String getToken(String email, TokenType tokenType) {
         return (String) redisTemplate.opsForValue().get(tokenType + "_" + email);
-
     }
 }
