+++ backend/src/main/java/com/example/fintrack/user/UserService.java
@@ -1,17 +1,27 @@
 package com.example.fintrack.user;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.function.Supplier;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserService implements UserDetailsService {
+
+    private final UserRepository userRepository;
     private final Supplier<User> principalSupplier;
 
     public String test(){
         User user = principalSupplier.get();
         return user.getEmail();
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return userRepository.findUserByEmail(username).orElseThrow(() -> new UsernameNotFoundException(username));
     }
 }
